pyproject.toml
[project]
name = "simpyvis"
version = "0.1.0"
description = ""
authors = [
    {name = "nikoliwilliams32",email = "nikoliwilliams32@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "simpy (>=4.1.1,<5.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "dash (>=3.2.0,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "numpy (<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
